// HourOfCode_Conny/bookstore-conny/apps/bookstore-conny/src/app/core/services/api/SwaggerWebAPI.ts
export class BookDto implements IBookDto {
    id?: number | null;
    name?: string | null;
    author?: string | null;
    year?: number | null;
    genres?: GenreDto[] | null;
    isLiked?: boolean| null;

    constructor(data?: IBookDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["id"] !== undefined ? data["id"] : <any>null;
            this.name = data["name"] !== undefined ? data["name"] : <any>null;
            this.author = data["author"] !== undefined ? data["author"] : <any>null;
            this.year = data["year"] !== undefined ? data["year"] : <any>null;
            if (data["genres"] && data["genres"].constructor === Array) {
                this.genres = [] as any;
                for (let item of data["genres"])
                    this.genres!.push(GenreDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): BookDto {
        data = typeof data === 'object' ? data : {};
        let result = new BookDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        data["author"] = this.author !== undefined ? this.author : <any>null;
        data["year"] = this.year !== undefined ? this.year : <any>null;
        if (this.genres && this.genres.constructor === Array) {
            data["genres"] = [];
            for (let item of this.genres)
                data["genres"].push(item.toJSON());
        }
        return data;
    }
}

export interface IBookDto {
    id?: number | null;
    name?: string | null;
    author?: string | null;
    year?: number | null;
    genres?: GenreDto[] | null;
}

export class GenreDto implements IGenreDto {
    id?: number | null;
    name?: string | null;

    constructor(data?: IGenreDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(data?: any) {
        if (data) {
            this.id = data["id"] !== undefined ? data["id"] : <any>null;
            this.name = data["name"] !== undefined ? data["name"] : <any>null;
        }
    }

    static fromJS(data: any): GenreDto {
        data = typeof data === 'object' ? data : {};
        let result = new GenreDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : <any>null;
        data["name"] = this.name !== undefined ? this.name : <any>null;
        return data;
    }
}

export interface IGenreDto {
    id?: number | null;
    name?: string | null;
}